Further Optimization
I optimized all of the pngs across the whole website. This is an important step in improving the SEO of the site. The next step is for me to integrate the new navigation bar that I have been working on. From there I will begin to create media queries site wide for mobile optimization.